Lesson4: homework edited

+++ src/main/java/lv/javaguru/student_artjomsb_homework/homework4/level_7_senior/Calculator.java
@@ -13,7 +13,7 @@ class Calculator {
         return firstNumber * secondNumber;
     }
 
-    public int div(int firstNumber, int secondNumber) {
+    public double div(int firstNumber, int secondNumber) {
         return firstNumber / secondNumber;
     }
